Remove AWSLab dependency and use classes from v2.0 AWS lambda library
