+++ spring-cloud-function-adapters/spring-cloud-function-adapter-aws/pom.xml
@@ -20,7 +20,7 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 		<java.version>1.8</java.version>
-		<aws-lambda-events.version>1.2.1</aws-lambda-events.version>
+		<aws-lambda-events.version>2.0.2</aws-lambda-events.version>
 	</properties>
 
 	<dependencies>
@@ -50,12 +50,6 @@
 			<groupId>com.amazonaws</groupId>
 			<artifactId>aws-lambda-java-events</artifactId>
 			<version>${aws-lambda-events.version}</version>
-			<scope>provided</scope>
-		</dependency>
-		<dependency>
-			<groupId>com.amazonaws.serverless</groupId>
-			<artifactId>aws-serverless-java-container-core</artifactId>
-			<version>0.8</version>
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
