+++ spring-cloud-function-samples/function-sample-aws/pom.xml
@@ -23,7 +23,7 @@
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 		<java.version>1.8</java.version>
 		<wrapper.version>1.0.9.RELEASE</wrapper.version>
-		<aws-lambda-events.version>1.2.1</aws-lambda-events.version>
+		<aws-lambda-events.version>2.0.2</aws-lambda-events.version>
 		<reactor.version>3.1.2.RELEASE</reactor.version>
 		<spring-cloud-function.version>1.0.0.BUILD-SNAPSHOT</spring-cloud-function.version>
 		<spring-cloud-stream-servlet.version>1.0.0.BUILD-SNAPSHOT</spring-cloud-stream-servlet.version>
